ai-backend/requirements.txt
pinecone
chromadb
numpy
pandas
python-dotenv
langchain
langchain-google-genai
google-generativeai
python-dotenv
langdetect
mtranslate